% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{calculate_convex_hull}
\alias{calculate_convex_hull}
\title{Calculate Convex Hulls for one variable}
\usage{
calculate_convex_hull(data, var_name, impute_method = c("mean", "median"))
}
\arguments{
\item{data}{Data frame containing the 'batch', 'order' and variable
'value' columns.}

\item{var_name}{Name of the variable to calculate convex hull for.}

\item{impute_method}{One of "mean" or "median".}
}
\value{
A list of dataframes of convex hull.
}
\description{
Calculate Convex Hulls for one variable
}
