% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{calculate_convex_indicators}
\alias{calculate_convex_indicators}
\title{Calculate the intra/inter batch dispersion indicators and their ratio on
convex hulls of a single variable.}
\usage{
calculate_convex_indicators(hull_data_list, var_name)
}
\arguments{
\item{hull_data_list}{list of data frames of convex hulls.}

\item{var_name}{name of the variable.}
}
\value{
A data frame with the indicators values.
}
\description{
Calculate the intra/inter batch dispersion indicators and their ratio on
convex hulls of a single variable.
}
