% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm_function.R
\name{compute_individual_global}
\alias{compute_individual_global}
\title{Computes Integrated Covariance Mahalanobis (ICM) distances of all
individuals in PCA-reduced space, against their global barycenter reference.}
\usage{
compute_individual_global(pc_data, batch_col)
}
\arguments{
\item{pc_data}{PCA-reduced data frame.}

\item{batch_col}{Name of the column representing batch or group.}
}
\value{
A data frame with Mahalanobis distances for each individual against
the global barycenter.
}
\description{
Computes Integrated Covariance Mahalanobis (ICM) distances of all
individuals in PCA-reduced space, against their global barycenter reference.
}
