% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm_function.R
\name{compute_intra}
\alias{compute_intra}
\title{Computes Integrated Covariance Mahalanobis (ICM) mean distances
within each batch in PCA-reduced space, using median and mean for center
references.}
\usage{
compute_intra(pc_data, batch_col)
}
\arguments{
\item{pc_data}{PCA-reduced data frame.}

\item{batch_col}{Name of the column representing batch or group.}
}
\value{
A data frame with Mahalanobis distances mean for each batch.
}
\description{
Computes Integrated Covariance Mahalanobis (ICM) mean distances
within each batch in PCA-reduced space, using median and mean for center
references.
}
