% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{compute_intra_batch_dispersion}
\alias{compute_intra_batch_dispersion}
\title{Calculate the intra batch dispersion indicator on convex hulls of a single
variable}
\usage{
compute_intra_batch_dispersion(hull_data_shoelace_list)
}
\arguments{
\item{hull_data_shoelace_list}{named list of convex hulls data frames with an additional column of shoelace core
values, for each batch.}
}
\value{
value of intra batch dispersion.
}
\description{
Calculate the intra batch dispersion indicator on convex hulls of a single
variable
}
