% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{compute_ratio}
\alias{compute_ratio}
\title{Calculate the intra/inter batch dispersion ratio indicator on convex hulls
of a single variable.}
\usage{
compute_ratio(intraB_disp, interB_disp)
}
\arguments{
\item{intraB_disp}{value of intra batch dispersion indicator.}

\item{interB_disp}{value of inter batch dispersion indicator.}
}
\value{
value of intra/inter batch dispersion ratio.
}
\description{
Calculate the intra/inter batch dispersion ratio indicator on convex hulls
of a single variable.
}
