% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{compute_shoelace_core}
\alias{compute_shoelace_core}
\title{Compute the shoelace core for convex hulls of a single variable}
\usage{
compute_shoelace_core(hull_data_list)
}
\arguments{
\item{hull_data_list}{named list of data frames of convex hulls, for each batch.}
}
\value{
named list of dataframes of convex hull concatenated with a column of shoelace core values, for each batch.
}
\description{
Compute the shoelace core for convex hulls of a single variable
}
