% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{hull_data_list_check}
\alias{hull_data_list_check}
\title{Function to check if hull_data_list is a valid list of data frames}
\usage{
hull_data_list_check(hull_data_list, name)
}
\arguments{
\item{hull_data_list}{List of data frames representing convex hulls.}

\item{name}{Name of the hull_data_list for error messages.}
}
\value{
None. The function raises an error if the checks fail.
}
\description{
Function to check if hull_data_list is a valid list of data frames
}
