% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{plot_all_convex_hulls}
\alias{plot_all_convex_hulls}
\title{Plot all convex hulls for each variable in a PDF file.}
\usage{
plot_all_convex_hulls(
  target_file_path,
  convex_analysis_res,
  show_points,
  mode = c("global", "batchwise")
)
}
\arguments{
\item{target_file_path}{Path to the output PDF file.}

\item{convex_analysis_res}{Result of the convex analysis containing data,
convex hulls and indicators.}

\item{show_points}{Boolean indicating whether to show points in the plot.}

\item{mode}{Mode of the analysis, either "global" or "batchwise".}
}
\value{
None. The function saves the plots to a PDF file.
}
\description{
Plot all convex hulls for each variable in a PDF file.
}
