% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{plot_convex_hull}
\alias{plot_convex_hull}
\title{Plot the convex hulls of a single variable.}
\usage{
plot_convex_hull(
  data,
  hull_data_list,
  var_name,
  show_points,
  label_prefix,
  indicators
)
}
\arguments{
\item{data}{Data frame containing the batch, order and variable
value columns.}

\item{hull_data_list}{List of data frames of convex hulls.}

\item{var_name}{Name of the variable.}

\item{show_points}{Boolean indicating whether to show points.}

\item{label_prefix}{Prefix for the plot title.}

\item{indicators}{Data frame with the indicators values.}
}
\value{
A ggplot object.
}
\description{
Plot the convex hulls of a single variable.
}
