% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm_function.R
\name{save_icm_distances_csv}
\alias{save_icm_distances_csv}
\title{Save ICM Distances to CSV Files}
\usage{
save_icm_distances_csv(distances, folder_path, prefix = "ICM")
}
\arguments{
\item{distances}{A list containing data.frames of distances (result from
\code{compute_icm_distances})}

\item{folder_path}{Path to the folder where files will be saved.}

\item{prefix}{Prefix for the output file names.}
}
\value{
None. Saves files to folder_path.
}
\description{
Save ICM Distances to CSV Files
}
