% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_function.R
\name{single_variable_df_check}
\alias{single_variable_df_check}
\title{Function to check if a single variable data frame is valid}
\usage{
single_variable_df_check(df, name)
}
\arguments{
\item{df}{Data frame containing 'batch', 'order', and 'value' columns.}

\item{name}{Name of the data frame for error messages.}
}
\value{
None. The function raises an error if the checks fail.
}
\description{
Function to check if a single variable data frame is valid
}
